+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.0.0] - 2024-12-12
+
+Breaking change in methods GeodeticCalculatorFactory and BearingCalculator
+
+### Changed
+
+Changed method signature on
+
+```
+GeodeticCalculatorFactory.createGeodeticCalculator
+BearingCalculator.calculateBearing
+```
+
+Removed srid parameter, these calculations always use WGS84 coordinate reference system.<br>
+Changed FractionAndDistanceCalculator always uses WGS84 coordinate reference system.
+
+### Added
+
+Validation on a coordinate reference system in FractionAndDistanceCalculator on input line strings.
+If no coordinate reference system is provided, it assumes WGS84.
+
 ## [3.2.0] - 2024-10-31
 
 ### Added